@@ -1,18 +1,18 @@
 import React       from 'react'
-import PropTypes   from 'prop-types'
 import FInputNumberBase from './FInputNumberBase.mjs'
 import useFloatProps from './useFloatProps.mjs'
 import {FLT_FLOAT_ES, FLT_FLOAT_EN} from '../../helpers/float/floatFilters.mjs'
 
-const FInputFloat = (props) => {
+const FInputFloat = ({decimalSign = '.', icon= 'dollar', ...props}) => {
 
-  const t= useFloatProps(props.decimalSign)
+  const t= useFloatProps(decimalSign)
 
   return (
     <FInputNumberBase {...
         {...props,
+          icon,
          inputFilter: 
-           props.decimalSign==','
+           decimalSign==','
           ? FLT_FLOAT_ES
           : FLT_FLOAT_EN,
          t
@@ -20,18 +20,5 @@ const FInputFloat = (props) => {
       }/>
   )
  }
-
-FInputFloat.propTypes = {
-  ...FInputNumberBase.propTypes,
-  decimals    : PropTypes.number,
-  decimalSign : PropTypes.oneOf([',', '.']),
-}
-
-FInputFloat.defaultProps = {
-  ...FInputNumberBase.defaultProps,
-  icon: 'dollar',
-  decimals: undefined,
-  decimalSign: '.'
-}
 
 export default FInputFloat
