@@ -3,7 +3,7 @@ import useIntProps from './useIntProps.mjs'
 import FInputNumberBase from './FInputNumberBase.mjs'
 
 
-const FInputUInt = (props) => {
+const FInputUInt = ({icon= 'dice', ...props}) => {
  
   const t= useIntProps(true)
 
@@ -11,18 +11,12 @@ const FInputUInt = (props) => {
 
     <FInputNumberBase {...
         {...props,
+          icon,
           inputFilter: 'uint',
           t
         }
       }/>
   )
-}
-
-FInputUInt.propTypes = FInputNumberBase.propTypes
-
-FInputUInt.defaultProps = {
-  ...FInputNumberBase.defaultProps,
-  icon: 'dice'
 }
 
 export default FInputUInt
