@@ -1,20 +1,11 @@
 import React      from 'react'
 import FInputText from './FInputText.mjs'
 
-const FInputTextArea = (props) =>
+const FInputTextArea = ({icon= 'text', ...props}) =>
 
   <FInputText {...props}
-                inputType = {"textarea"}/>
-
-FInputTextArea.propTypes = FInputText.propTypes
-
-delete FInputTextArea.propTypes['inputType']
-
-FInputTextArea.defaultProps = {
-  ...FInputText.defaultProps,
-  icon: 'text'
-}
-
+    icon = {icon}
+    inputType = {"textarea"}/>
 
 
 export default FInputTextArea
